+++ TransitCity/Utility/Units/Area.cs
@@ -4,11 +4,11 @@ namespace Utility.Units
 {
     public class Area
     {
-        private readonly float _squareMeters;
+        private readonly double _squareMeters;
 
-        internal Area(float squareMeters)
+        internal Area(double squareMeters)
         {
-            if (squareMeters < 0f)
+            if (squareMeters < 0.0)
             {
                 throw new ArgumentOutOfRangeException(nameof(squareMeters));
             }
@@ -16,9 +16,9 @@ namespace Utility.Units
             _squareMeters = squareMeters;
         }
 
-        public float SquareMeters => _squareMeters;
+        public double SquareMeters => _squareMeters;
 
-        public float SquareKilometers => _squareMeters / 1000000f;
+        public double SquareKilometers => _squareMeters / 1000000.0;
 
         public override string ToString()
         {
@@ -29,11 +29,11 @@ namespace Utility.Units
 
         public static Area operator -(Area a1, Area a2) => new Area(a1.SquareMeters - a2.SquareMeters);
 
-        public static Area operator *(Area a, float f) => new Area(a.SquareMeters * f);
+        public static Area operator *(Area a, double f) => new Area(a.SquareMeters * f);
 
-        public static Area operator *(float f, Area a) => a * f;
+        public static Area operator *(double f, Area a) => a * f;
 
-        public static Area operator /(Area a, float f) => new Area(a.SquareMeters / f);
+        public static Area operator /(Area a, double f) => new Area(a.SquareMeters / f);
 
         public static bool operator <(Area a1, Area a2) => a1.SquareMeters < a2.SquareMeters;
 
@@ -43,8 +43,8 @@ namespace Utility.Units
 
         public static bool operator >=(Area a1, Area a2) => a1.SquareMeters >= a2.SquareMeters;
 
-        public static Area FromSqaureMeters(float squareMeters) => new Area(squareMeters);
+        public static Area FromSqaureMeters(double squareMeters) => new Area(squareMeters);
 
-        public static Area FromSquareKilometers(float squareKilometers) => new Area(squareKilometers * 1000000);
+        public static Area FromSquareKilometers(double squareKilometers) => new Area(squareKilometers * 1000000.0);
     }
 }
