+++ TransitCity/Utility/Units/Distance.cs
@@ -4,11 +4,11 @@ namespace Utility.Units
 {
     public class Distance
     {
-        private readonly float _meters;
+        private readonly double _meters;
 
-        internal Distance(float meters)
+        internal Distance(double meters)
         {
-            if (meters < 0f)
+            if (meters < 0.0)
             {
                 throw new ArgumentOutOfRangeException(nameof(meters));
             }
@@ -16,9 +16,9 @@ namespace Utility.Units
             _meters = meters;
         } 
 
-        public float Meters => _meters;
+        public double Meters => _meters;
 
-        public float Kilometers => _meters / 1000;
+        public double Kilometers => _meters / 1000.0;
 
         public override string ToString()
         {
@@ -29,13 +29,13 @@ namespace Utility.Units
 
         public static Distance operator -(Distance d1, Distance d2) => new Distance(d1.Meters - d2.Meters);
 
-        public static Distance operator *(Distance d, float f) => new Distance(d.Meters * f);
+        public static Distance operator *(Distance d, double f) => new Distance(d.Meters * f);
 
-        public static Distance operator *(float f, Distance d) => d * f;
+        public static Distance operator *(double f, Distance d) => d * f;
 
         public static Area operator *(Distance d1, Distance d2) => new Area(d1.Meters * d2.Meters);
 
-        public static Distance operator /(Distance d, float f) => new Distance(d.Meters / f);
+        public static Distance operator /(Distance d, double f) => new Distance(d.Meters / f);
 
         public static Speed operator /(Distance d, Duration t) => new Speed(d.Meters / t.Seconds);
 
@@ -51,8 +51,8 @@ namespace Utility.Units
 
         public static bool operator >=(Distance d1, Distance d2) => d1.Meters >= d2.Meters;
 
-        public static Distance FromMeters(float meters) => new Distance(meters);
+        public static Distance FromMeters(double meters) => new Distance(meters);
 
-        public static Distance FromKilometers(float kilometers) => new Distance(kilometers * 1000);
+        public static Distance FromKilometers(double kilometers) => new Distance(kilometers * 1000.0);
     }
 }
